The changes were made in the following methods of the file -- • getDataPoint • getRatio • Main
Where in "getDataPoint" method I have modified to compute the right stock price, changed how `price` is computed by using the formula: (bid_price+ask_price) / 2.

Where in "getRatio" method I have modified  the return value to the ratio of stock price_a to stock price_b,
also I have added an if condition for the case where price_b could be zero.

In "Main" method, I created a prices dictionary to store the stock prices. The key is the stock name and the value as the price. Then used this prices dictionary at the end to pass in the right values in the getRatio function.